Rename the `versions` mode to `history`, and delete the old history template (D243)
Two modes were reading "History" side by side on the same registry keys.
`versions` — the git commit spine that lists a target's log, materialises any
commit as a browsable snapshot and reverts by committing on top — is what a
user means by a file's history, and `versions` is not the word they reach for.
The other one, `templates/history/` from D96, was a read-only inspector over a
file's `<ext>.json` sidecar: it parsed the sidecar in the browser (no `.py` at
all), validated each top-level key against an inline per-key schema, and drew
claude sessions, bookmark history, the last-session restore card and annotate
comments as one merged timeline. D237 had already removed its one interactive
affordance — the comment deep link into `annotate` — leaving a viewer for a
store the surfaces that own it already show in context.

So the word goes to the mode that earns it and the other view goes away:

* `templates/versions/` -> `templates/history/`, `versions.py` -> `history.py`
  (a template's python module is reached by relative path from its own
  template.html, so the module renames with the folder).
* `templates/history/` (icon + the ~1040-line sidecar template) and
  `tests/test_history_template.py` deleted; `tests/test_versions_template.py`
  takes over the name.
* `".*.json"` falls back to `["tree", "code"]` — the same bytes, without
  pretending to be a history. `.html` and `.parquet` lose the duplicate entry
  and keep the surviving mode in the slot `versions` held.

The collision this closes is exactly the one `mode-labels.test.ts` was written
for; that guard stays, because it catches the shape of the bug rather than this
pair, and its comment now tells the new story.

Two things deliberately do NOT follow the rename, each with a comment at its
definition saying why. The snapshot cache `~/.fused-render/app-versions/` keeps
its spelling: it is a materialisation directory no user sees or types, and
renaming it would strand every snapshot already extracted on disk behind a
prefix nothing looks under (`server/mount.py`'s prefix check is the other half
of that contract). And no alias is shipped for `?_mode=versions` — a stale deep
link falls back to the target's default view, which is PT-9's ordinary
behaviour, and a permanent alias would keep the dead token alive in every
registry a user forks.

SPEC §24 is kept as a tombstone so the HV-* ids other sections still cite
resolve to an explanation rather than to nothing.
