+++ fused_render/templates/registry.json
@@ -4,7 +4,6 @@
     "structure",
     "h3",
     "claude",
-    "versions",
     "history",
     "geometry_editor"
   ],
@@ -13,14 +12,14 @@
     "excel",
     "code",
     "claude",
-    "versions",
+    "history",
     "reader"
   ],
   ".tsv": [
     "duckdb",
     "code",
     "claude",
-    "versions",
+    "history",
     "reader"
   ],
   ".xlsx": [
@@ -37,7 +36,6 @@
     "slides"
   ],
   ".*.json": [
-    "history",
     "tree",
     "code"
   ],
@@ -46,20 +44,20 @@
     "code",
     "duckdb",
     "claude",
-    "versions",
+    "history",
     "reader"
   ],
   ".jsonl": [
     "code",
     "duckdb",
     "claude",
-    "versions",
+    "history",
     "reader"
   ],
   ".ipynb": [
     "notebook",
     "claude",
-    "versions"
+    "history"
   ],
   ".calls.jsonl": [
     "log_studio",
@@ -77,7 +75,7 @@
     "code",
     "duckdb",
     "claude",
-    "versions",
+    "history",
     "reader"
   ],
   ".csv.gz": [
@@ -126,70 +124,70 @@
     "tree",
     "code",
     "claude",
-    "versions",
+    "history",
     "geometry_editor"
   ],
   ".md": [
     "markdown",
     "code",
     "claude",
-    "versions",
+    "history",
     "reader"
   ],
   ".markdown": [
     "markdown",
     "code",
     "claude",
-    "versions",
+    "history",
     "reader"
   ],
   ".tex": [
     "latex",
     "code",
     "claude",
-    "versions",
+    "history",
     "reader"
   ],
   ".ltx": [
     "latex",
     "code",
     "claude",
-    "versions",
+    "history",
     "reader"
   ],
   ".latex": [
     "latex",
     "code",
     "claude",
-    "versions",
+    "history",
     "reader"
   ],
   ".svg": [
     "image",
     "code",
     "claude",
-    "versions"
+    "history"
   ],
   ".png": [
     "image",
     "photos",
     "pano",
     "claude",
-    "versions"
+    "history"
   ],
   ".jpg": [
     "image",
     "photos",
     "pano",
     "claude",
-    "versions"
+    "history"
   ],
   ".jpeg": [
     "image",
     "photos",
     "pano",
     "claude",
-    "versions"
+    "history"
   ],
   ".gif": [
     "image",
@@ -251,180 +249,179 @@
     "code",
     "api",
     "claude",
-    "versions",
+    "history",
     "reader"
   ],
   ".js": [
     "code",
     "claude",
-    "versions",
+    "history",
     "reader"
   ],
   ".ts": [
     "code",
     "claude",
-    "versions",
+    "history",
     "reader"
   ],
   ".tsx": [
     "code",
     "claude",
-    "versions",
+    "history",
     "reader"
   ],
   ".jsx": [
     "code",
     "claude",
-    "versions",
+    "history",
     "reader"
   ],
   ".cjs": [
     "code",
     "claude",
-    "versions",
+    "history",
     "reader"
   ],
   ".mjs": [
     "code",
     "claude",
-    "versions",
+    "history",
     "reader"
   ],
   ".cts": [
     "code",
     "claude",
-    "versions",
+    "history",
     "reader"
   ],
   ".mts": [
     "code",
     "claude",
-    "versions",
+    "history",
     "reader"
   ],
   ".sh": [
     "code",
     "claude",
-    "versions",
+    "history",
     "reader"
   ],
   ".zsh": [
     "code",
     "claude",
-    "versions",
+    "history",
     "reader"
   ],
   ".fish": [
     "code",
     "claude",
-    "versions",
+    "history",
     "reader"
   ],
   ".ps1": [
     "code",
     "claude",
-    "versions",
+    "history",
     "reader"
   ],
   ".csh": [
     "code",
     "claude",
-    "versions",
+    "history",
     "reader"
   ],
   ".zsh-theme": [
     "code",
     "claude",
-    "versions",
+    "history",
     "reader"
   ],
   ".vim": [
     "code",
     "claude",
-    "versions",
+    "history",
     "reader"
   ],
   ".yaml": [
     "code",
     "claude",
-    "versions",
+    "history",
     "reader"
   ],
   ".yml": [
     "code",
     "claude",
-    "versions",
+    "history",
     "reader"
   ],
   ".toml": [
     "canvas",
     "code",
     "claude",
-    "versions",
+    "history",
     "reader"
   ],
   ".ini": [
     "code",
     "claude",
-    "versions",
+    "history",
     "reader"
   ],
   ".cfg": [
     "code",
     "claude",
-    "versions",
+    "history",
     "reader"
   ],
   ".conf": [
     "code",
     "claude",
-    "versions",
+    "history",
     "reader"
   ],
   ".tf": [
     "code",
     "claude",
-    "versions",
+    "history",
     "reader"
   ],
   ".hcl": [
     "code",
     "claude",
-    "versions",
+    "history",
     "reader"
   ],
   ".css": [
     "code",
     "claude",
-    "versions",
+    "history",
     "reader"
   ],
   ".txt": [
     "code",
     "claude",
-    "versions",
+    "history",
     "reader"
   ],
   ".log": [
     "log_studio",
     "code",
     "claude",
-    "versions",
+    "history",
     "reader"
   ],
   ".html": [
     "_render",
     "code",
     "claude",
-    "versions",
-    "reader",
-    "history"
+    "history",
+    "reader"
   ],
   ".htm": [
     "_render",
     "code",
     "claude",
-    "versions",
+    "history",
     "reader"
   ],
   ".tif": [
@@ -450,7 +447,7 @@
     "_listing",
     "app",
     "claude",
-    "versions",
+    "history",
     "git",
     "graph",
     "zarr_aoi"
@@ -614,7 +611,7 @@
     "plist",
     "code",
     "claude",
-    "versions",
+    "history",
     "reader"
   ]
 }
