+++ tests/_theme_sources.py
@@ -46,7 +46,6 @@ TIER_ONE_TEMPLATES = (
     "text",
     "tree",
     "vector",
-    "versions",
     "xlsx",
     "zip",
 )
