+++ tests/test_canvas_template.py
@@ -108,7 +108,7 @@ def _write_canvas(dir_path, name="canvas.toml", body=CANVAS_TOML):
 def test_canvas_toml_gets_canvas_mode_first(tmp_path):
     p = _write_canvas(tmp_path / "cv")
     m, error = modes(str(p))
-    assert m == ["canvas", "code", "claude", "versions", "reader"]
+    assert m == ["canvas", "code", "claude", "history", "reader"]
     assert error is None
     entries, _ = _server_templates._templates_for(str(p), False)
     assert entries[0]["path"].endswith("canvas/template.html")
@@ -122,7 +122,7 @@ def test_canvas_toml_gets_canvas_mode_first(tmp_path):
     assert "conditional" not in by_mode["code"]
     # The gated file-side pair (D235) is marked too — both ship a condition.py.
     assert by_mode["claude"].get("conditional") is True
-    assert by_mode["versions"].get("conditional") is True
+    assert by_mode["history"].get("conditional") is True
     # … and the background verdict allows a genuine canvas.
     allowed, err = canvas_verdict(str(p))
     assert allowed is True
@@ -134,7 +134,7 @@ def test_plain_toml_denied_canvas_mode(tmp_path):
     p.write_text("[tool.black]\nline-length = 88\n")
     m, error = modes(str(p))
     # stat still lists canvas (marked conditional, gate not run at stat time)
-    assert m == ["canvas", "code", "claude", "versions", "reader"]
+    assert m == ["canvas", "code", "claude", "history", "reader"]
     assert error is None
     allowed, err = canvas_verdict(str(p))
     assert allowed is False  # basename pre-check denies it
